requirements.txt
chromadb==1.4.1
ollama
pypdf
python-docx
requests
sentence-transformers
streamlit
